# swagger.yaml
openapi: 3.0.3
info:
  title: Clinic-AI Intake Assistant API
  description: |
    AI-powered clinical intake system for small and mid-sized clinics.

    This API provides endpoints for patient registration and intake management,
    including AI-powered adaptive questioning for Step-01 functionality.

    ## Features
    - Patient registration with automatic ID generation
    - AI-powered adaptive questioning based on disease/complaint
    - Intake session management with question tracking
    - Support for multiple diseases (Hypertension, Diabetes, Chest Pain, Fever, Cough, Headache, Back Pain)
    - Maximum 12 questions per intake with early stopping capability

    ## Authentication
    Currently no authentication is required for development.

    ## Rate Limiting
    No rate limiting is currently implemented.
  version: 0.1.0
  contact:
    name: Clinic-AI Support
    email: support@clinicai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.clinicai.com
    description: Production server
...
paths:
...
  /patients/:
    post:
      tags:
        - patients
      summary: Register a new patient
      description: |
        Register a new patient and start an intake session.

        This endpoint:
        1. Validates patient data (name, mobile, age, disease)
        2. Generates a unique patient ID in format: CLINIC01_{name}_{mobile}
        3. Creates a new visit with ID in format: CONSULT-YYYYMMDD-XXX
        4. Generates the first AI-powered question based on the disease/complaint
        5. Returns patient_id, visit_id, and the first question

        ## Supported Diseases
        - Hypertension
        - Diabetes
        - Chest Pain
        - Fever
        - Cough
        - Headache
        - Back Pain
      operationId: register_patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPatientRequest"
            examples:
              hypertension_patient:
                summary: Hypertension Patient
                value:
                  name: "John Doe"
                  mobile: "9123456789"
                  age: 35
                  disease: "Hypertension"
              diabetes_patient:
                summary: Diabetes Patient
                value:
                  name: "Jane Smith"
                  mobile: "9876543210"
                  age: 45
                  disease: "Diabetes"
              chest_pain_patient:
                summary: Chest Pain Patient
                value:
                  name: "Bob Johnson"
                  mobile: "5551234567"
                  age: 60
                  disease: "Chest Pain"
      responses:
        "201":
          description: Patient registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterPatientResponse"
              examples:
                success_response:
                  summary: Successful Registration
                  value:
                    patient_id: "CLINIC01_JOHNDOE_9123456789"
                    visit_id: "CONSULT-20250103-001"
                    first_question: "What specific symptoms are you experiencing with your hypertension?"
                    message: "Patient registered successfully. Intake session started."
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validation_error:
                  summary: Validation Error
                  value:
                    error: "VALIDATION_ERROR"
                    message: "Name cannot be empty"
                    details: {}
        "409":
          description: Duplicate patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicate_patient:
                  summary: Duplicate Patient
                  value:
                    error: "DUPLICATE_PATIENT"
                    message: "Patient with ID 'CLINIC01_JOHNDOE_9123456789' already exists"
                    details:
                      patient_id: "CLINIC01_JOHNDOE_9123456789"
        "422":
          description: Invalid disease
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_disease:
                  summary: Invalid Disease
                  value:
                    error: "INVALID_DISEASE"
                    message: "Invalid disease/complaint: InvalidDisease"
                    details:
                      disease: "InvalidDisease"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/consultations/answer:
    post:
      tags:
        - patients
      summary: Answer an intake question
      description: |
        Answer an intake question and get the next question or completion status.

        This endpoint:
        1. Validates the answer and finds the patient/visit
        2. Adds the answer to the intake session
        3. Uses AI to determine if sufficient information has been collected
        4. Generates the next adaptive question or completes the intake
        5. Returns the next question or completion status

        ## Question Flow
        - Maximum 12 questions per intake session
        - AI can stop early if sufficient information is collected
        - No duplicate questions are asked
        - Questions are adaptive based on previous answers
      operationId: answer_intake_question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerIntakeRequest"
            examples:
              first_answer:
                summary: First Answer
                value:
                  patient_id: "CLINIC01_JOHNDOE_9123456789"
                  visit_id: "CONSULT-20250103-001"
                  answer: "I have been experiencing high blood pressure for 3 months"
              follow_up_answer:
                summary: Follow-up Answer
                value:
                  patient_id: "CLINIC01_JOHNDOE_9123456789"
                  visit_id: "CONSULT-20250103-001"
                  answer: "The headaches are severe, about 8 out of 10 on the pain scale"
      responses:
        "200":
          description: Answer processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerIntakeResponse"
              examples:
                next_question:
                  summary: Next Question Available
                  value:
                    next_question: "How long have you been experiencing these symptoms?"
                    is_complete: false
                    question_count: 1
                    max_questions: 12
                    message: "Question 2 of 12"
                intake_complete:
                  summary: Intake Completed
                  value:
                    next_question: null
                    is_complete: true
                    question_count: 5
                    max_questions: 12
                    message: "Intake completed successfully. Ready for next step."
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Patient or visit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                patient_not_found:
                  summary: Patient Not Found
                  value:
                    error: "PATIENT_NOT_FOUND"
                    message: "Patient with ID 'CLINIC01_INVALID_1234567890' not found"
                    details:
                      patient_id: "CLINIC01_INVALID_1234567890"
                visit_not_found:
                  summary: Visit Not Found
                  value:
                    error: "VISIT_NOT_FOUND"
                    message: "Visit with ID 'CONSULT-20250103-999' not found"
                    details:
                      visit_id: "CONSULT-20250103-999"
        "409":
          description: Intake already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                intake_completed:
                  summary: Intake Already Completed
                  value:
                    error: "INTAKE_ALREADY_COMPLETED"
                    message: "Intake already completed for visit: CONSULT-20250103-001"
                    details:
                      visit_id: "CONSULT-20250103-001"
        "422":
          description: Question limit exceeded or duplicate question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                question_limit:
                  summary: Question Limit Exceeded
                  value:
                    error: "QUESTION_LIMIT_EXCEEDED"
                    message: "Question limit exceeded. Current: 12, Max: 12"
                    details:
                      current_count: 12
                      max_count: 12
                duplicate_question:
                  summary: Duplicate Question
                  value:
                    error: "DUPLICATE_QUESTION"
                    message: "Duplicate question detected: What symptoms are you experiencing?"
                    details:
                      question: "What symptoms are you experiencing?"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    RegisterPatientRequest:
      type: object
      required:
        - name
        - mobile
        - age
        - disease
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: Patient's full name
          example: "John Doe"
        mobile:
          type: string
          minLength: 10
          maxLength: 15
          pattern: '^\d{10,15}$'
          description: Mobile phone number (digits only)
          example: "9123456789"
        age:
          type: integer
          minimum: 0
          maximum: 120
          description: Patient's age in years
          example: 35
        disease:
          type: string
          enum:
            - "Hypertension"
            - "Diabetes"
            - "Chest Pain"
            - "Fever"
            - "Cough"
            - "Headache"
            - "Back Pain"
          description: Primary disease or complaint
          example: "Hypertension"
    RegisterPatientResponse:
      type: object
      required:
        - patient_id
        - visit_id
        - first_question
        - message
      properties:
        patient_id:
          type: string
          pattern: '^CLINIC01_[A-Z0-9]+_\d+$'
          description: Generated patient ID in format CLINIC01_{name}_{mobile}
          example: "CLINIC01_JOHNDOE_9123456789"
        visit_id:
          type: string
          pattern: '^CONSULT-\d{8}-\d{3}$'
          description: Generated visit ID in format CONSULT-YYYYMMDD-XXX
          example: "CONSULT-20250103-001"
        first_question:
          type: string
          description: AI-generated first question based on the disease
          example: "What specific symptoms are you experiencing with your hypertension?"
        message:
          type: string
          description: Success message
          example: "Patient registered successfully. Intake session started."
    AnswerIntakeRequest:
      type: object
      required:
        - patient_id
        - visit_id
        - answer
      properties:
        patient_id:
          type: string
          pattern: '^CLINIC01_[A-Z0-9]+_\d+$'
          description: Patient ID from registration
          example: "CLINIC01_JOHNDOE_9123456789"
        visit_id:
          type: string
          pattern: '^CONSULT-\d{8}-\d{3}$'
          description: Visit ID from registration
          example: "CONSULT-20250103-001"
        answer:
          type: string
          minLength: 1
          maxLength: 1000
          description: Patient's answer to the question
          example: "I have been experiencing high blood pressure for 3 months"
    AnswerIntakeResponse:
      type: object
      required:
        - next_question
        - is_complete
        - question_count
        - max_questions
        - message
      properties:
        next_question:
          type: string
          nullable: true
          description: Next question to ask (null if intake is complete)
          example: "How long have you been experiencing these symptoms?"
        is_complete:
          type: boolean
          description: Whether the intake session is complete
          example: false
        question_count:
          type: integer
          minimum: 0
          maximum: 12
          description: Current number of questions asked
          example: 1
        max_questions:
          type: integer
          description: Maximum number of questions allowed
          example: 12
        message:
          type: string
          description: Status message
          example: "Question 2 of 12"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Name cannot be empty"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          example:
            field: "name"
            value: ""
...
  securitySchemes:
...
security:
  # No security requirements for development
  - {}